requirements.txt
affine==2.4.0
chardet
dearpygui
GPUtil
imageio
joblib
jupyter
kornia==0.5.3
lpips
matplotlib
ninja
numba==0.57.1
numpy==1.24.3
opencv-contrib-python
opencv-python==4.7.0.72
packaging
pandas
Pillow==9.2.0
plotly==5.16.1
plyflatten==0.2.0
PyMCubes
pymp-pypi
pyproj==3.0.1
PyYAML
rasterio==1.2.10
rich
rpcm
scipy
srtm4>=1.2.1
tensorboardX==2.6.1
torch-ema==0.3
torchmetrics==0.11.4
tqdm
trimesh
utm==0.7.0
